% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleShrink.R
\name{simpleShrink}
\alias{simpleShrink}
\title{Simple shrink}
\usage{
simpleShrink(
  unpooled_estimates,
  pooled_estimate,
  unpooled_std_error,
  pooled_std_error
)
}
\arguments{
\item{unpooled_estimates}{A vector of unpooled point estimates.}

\item{pooled_estimate}{A value of pooled point esitmates}

\item{unpooled_std_error}{A vector of unpooled standard errors.}

\item{pooled_std_error}{A value of pooled standard error.}
}
\value{
A list of components:
\describe{
  \item{eb}{A vector of EB estimates.}
  \item{posterior_sd}{A vector of posterior standard deviations.}
  \item{pooled_weight}{A vector of shrinkage weights on the pooled estimate.}
  \item{unpooled_weight}{A vector of shrinkage weights on unpooled estimates.}
}
}
\description{
\code{simpleShrink} takes pooled and unpooled estimates as input and returns 
EB estimates and corresponding shrinkage weights.
}
